The Professor and University classes have been created with attributes, constructors, getters, and setters.
 The showInfo() method has been added to each class to display information on the console.

 A menu with action choices has been created in Main (while(true) loop).

 Menu item 1 has been implemented to add a professor by entering all the data and storing it in an ArrayList<Professor>.

 Menu item 2 has been implemented to add a university by entering all the data and storing it in an ArrayList<University>.

Menu item 3 is implemented — showing all objects, where all professors and universities are displayed using showInfo().

Menu item 4 is implemented — comparing objects, where:

the professor with the most experience is searched for,

the university with the least/oldest founding year is searched for,

it is checked that there are at least 2 objects in the lists.

Menu item 5 is implemented — exiting the program (break from the loop).

Fixed a problem with Scanner: after nextInt(), in.nextLine() is added to correctly read lines after numbers.

 ArrayLists are used to store multiple objects and allow for repeated additions.